Added slurm job script to game-of-life

+++ game-of-life/run.sh
@@ -1,10 +1,10 @@
 
 # make sure we have the correct arguments
-if [ "$#" != 5 ]
+if [ "$#" != 7 ]
 then
-  echo "We will specify the size of the game board, and provide a random seed that will be used to generate the same board for the serial and parallel versions. We will also need to specify how many generations to run the game for (-1 for indefinite):"
+  echo "We will specify the size of the game board, and provide a random seed that will be used to generate the same board for the serial and parallel versions. We will also need to specify how many generations to run the game for:"
   echo
-  echo "Usage: ${0} <rows> <columns> <seed> <generation> <number of processes>"
+  echo "Usage: ${0} <rows> <columns> <seed> <generation> <number of processes> <run serial? [y/n]> <run parallel? [y/n]>"
   exit
 fi
 
@@ -14,38 +14,50 @@ columns=$2
 seed=$3
 generation=$4
 numProcs=$5
+runSerial=$6
+runParallel=$7
 
 serialFile="serial-output.txt"
 parallelFile="parallel-output.txt"
 
 # make the serial and parallel versions
-echo "Making executables"
+# echo "Making executables"
 make clean
 make
-echo "Done making"
-echo
+# echo "Done making"
+# echo
 
 # run the serial version, pipe the output to the relevant file
-echo "Running serial"
-rm -f $serialFile
-./serial $rows $columns $seed $generation 
-echo "Done serial"
-echo
+if [ $runSerial == "y" ]
+then
+  # echo "Running serial"
+  rm -f $serialFile
+  ./serial $rows $columns $seed $generation 
+  # echo "Done serial"
+  echo
+fi
 
 # run the parallel version, pipe the output to the relevant file
-echo "Running parallel"
-rm -f $parallelFile
-mpirun -np $numProcs ./parallel $rows $columns $seed $generation 
-echo "Done parallel"
-echo
+if [ $runParallel == "y" ]
+then
+  # echo "Running parallel"
+  rm -f $parallelFile
+  mpirun -np $numProcs ./parallel $rows $columns $seed $generation 
+  # echo "Done parallel"
+  echo
+fi
 
 # compare the serial and parallel outputs to VERIFY
-DIFF=$(diff $serialFile $parallelFile)
-if [ "$DIFF" ]
-then 
-  echo "The serial and parallel outputs are different!"
-else
-  echo "The serial and parallel outputs are the same and correct!"
+# only compare if we want to check the parallel version
+if [ $runParallel == "y" ] 
+then
+  DIFF=$(diff $serialFile $parallelFile)
+  if [ "$DIFF" ]
+  then 
+    echo "The serial and parallel outputs are different!"
+  else
+    echo "The serial and parallel outputs are the same and correct!"
+  fi
 fi
 
 # clean up
